requirements.txt
netifaces
upnpclient
ifaddr
dpkt
requests
psutil
pcap-ct
